This PR reduces the amount of `string -> *regexp.Regexp -> string` in reference Regexps. This conversion (especially the compilation is quite expensive both in memory and CPU usage.

This PR tries to be as minimal as possible by ensuring that the same regexps are exported (even those that are not really exported but just used in the same package) but I would be happy to change it to anything else if bigger changes are desired.

Signed-off-by: Paul Cacheux <paul.cacheux@datadoghq.com>