@deleteriousEffect 2022-01-13
Looking at the code, it seems like it would be more efficient, but would you also mind providing a couple of benchmarks, so we can quantify the improvement?