@deleteriousEffect 2022-01-18
> Here is a repo with benchmarks around this PR https://github.com/paulcacheux/distrib-re-benchmark. Since it's benchmarking an init function it's a bit hacky but it works.

I really appreciate this, thank you!

For posterity, this is the output from my machine:

###### Previous
```
10:22 $ go test -bench=. -benchmem
goos: linux
goarch: amd64
pkg: example.com/previous
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkPrevious-16                1269           1586451 ns/op         1339721 B/op      11453 allocs/op
PASS
ok      example.com/previous    2.114s
```

###### This PR
```
10:21 $ go test -bench=. -benchmem
goos: linux
goarch: amd64
pkg: example.com/pr
cpu: Intel(R) Core(TM) i9-9980HK CPU @ 2.40GHz
BenchmarkPR-16              3660            822432 ns/op          485248 B/op       4501 allocs/op
PASS
ok      example.com/pr  3.047s
```